feat: Challenge 02 completed
